requirements.txt
langchain
langchain-community
openai
python-dotenv
